fix: Update spritesheet
A new prop was added in the Swampy Shores update. Attempting to asset swap the new spritesheet to the one in Doctorpus Assets will result in the PIXI.JS spritesheet loader throwing an error. This is due to a mismatch between the dimensions of the spritesheet and the dimensions specified in the accompanying JSON file.